ai_ad_agency/requirements.txt
# AI Ad Agency — Python Dependencies
# Install with: pip install -r requirements.txt
# Requires Python 3.11+

# ---- Core ----
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# ---- CLI ----
typer[all]>=0.9.0

# ---- LLM ----
openai>=1.12.0

# ---- HTTP / Networking ----
requests>=2.31.0
httpx>=0.27.0

# ---- Retry / Resilience ----
tenacity>=8.2.3

# ---- Image Processing ----
Pillow>=10.2.0
opencv-python-headless>=4.9.0

# ---- Media / FFmpeg ----
# FFmpeg must be installed system-wide: https://ffmpeg.org/download.html
# moviepy is optional — we prefer raw FFmpeg subprocess calls
# moviepy>=1.0.3

# ---- Data / Export ----
pandas>=2.1.0
numpy>=1.26.0

# ---- CLI UX ----
rich>=13.7.0

# ---- Database ----
# sqlite3 is built into Python — no extra package needed

# ---- Async (optional, for future async provider support) ----
aiohttp>=3.9.0
asyncio-throttle>=1.0.2

# ---- Testing ----
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# ---- Development (optional) ----
# black>=23.0.0
# ruff>=0.1.0
# mypy>=1.7.0
